update INTEGER fields to BIGINT to support larger integers

+++ src/sql/airdrops.sql
@@ -1,12 +1,12 @@
 DROP TABLE IF EXISTS airdrops;
 CREATE TABLE airdrops (
-    action_index      INTEGER UNSIGNED NOT NULL, -- Unique action index
-    tick_id           INTEGER UNSIGNED,          -- id of record in index_ticks
-    source_id         INTEGER UNSIGNED,          -- id of record in index_addresses table
-    list_action_index INTEGER UNSIGNED,          -- list action_index
+    action_index      BIGINT UNSIGNED NOT NULL, -- Unique action index
+    tick_id           BIGINT UNSIGNED,          -- id of record in index_ticks
+    source_id         BIGINT UNSIGNED,          -- id of record in index_addresses table
+    list_action_index BIGINT UNSIGNED,          -- list action_index
     amount            VARCHAR(250),              -- Amount of token in airdrop
-    memo_id           INTEGER UNSIGNED,          -- id of record in index_memos table 
-    status_id         INTEGER UNSIGNED           -- id of record in index_statuses table
+    memo_id           BIGINT UNSIGNED,          -- id of record in index_memos table 
+    status_id         BIGINT UNSIGNED           -- id of record in index_statuses table
 ) ENGINE=MyISAM DEFAULT CHARSET=utf8 COLLATE=utf8_general_ci;
 
 CREATE UNIQUE INDEX action_index      ON airdrops (action_index);
