+++ src/sql/balances.sql
@@ -1,8 +1,8 @@
 DROP TABLE IF EXISTS balances;
 CREATE TABLE balances (
-    id         INTEGER UNSIGNED NOT NULL AUTO_INCREMENT PRIMARY KEY,
-    address_id INTEGER UNSIGNED, -- id of record in index_addresses
-    tick_id    INTEGER UNSIGNED, -- id of record in index_tickers
+    id         BIGINT UNSIGNED NOT NULL AUTO_INCREMENT PRIMARY KEY,
+    address_id BIGINT UNSIGNED, -- id of record in index_addresses
+    tick_id    BIGINT UNSIGNED, -- id of record in index_tickers
     amount     VARCHAR(250)      -- AMOUNT of balance
 ) ENGINE=MyISAM DEFAULT CHARSET=utf8 COLLATE=utf8_general_ci;
