+++ src/sql/fees.sql
@@ -1,11 +1,11 @@
 DROP TABLE IF EXISTS fees;
 CREATE TABLE fees (
-    action_index   INTEGER UNSIGNED NOT NULL, -- Unique action index
-    source_id      INTEGER UNSIGNED,          -- id of record in index_addresses table
-    tick_id        INTEGER UNSIGNED,          -- id of record in index_tickers (default = GAS) 
+    action_index   BIGINT UNSIGNED NOT NULL, -- Unique action index
+    source_id      BIGINT UNSIGNED,          -- id of record in index_addresses table
+    tick_id        BIGINT UNSIGNED,          -- id of record in index_tickers (default = GAS) 
     amount         VARCHAR(250),              -- Amount of TICK
-    method         INTEGER UNSIGNED NOT NULL, -- FEE Payment Method (1=Destroy, 2=Donate)
-    destination_id INTEGER UNSIGNED           -- id of record in index_addresses table
+    method         BIGINT UNSIGNED NOT NULL, -- FEE Payment Method (1=Destroy, 2=Donate)
+    destination_id BIGINT UNSIGNED           -- id of record in index_addresses table
 ) ENGINE=MyISAM DEFAULT CHARSET=utf8 COLLATE=utf8_general_ci;
 
 CREATE UNIQUE INDEX action_index   ON fees (action_index);
