+++ src/sql/issues.sql
@@ -1,14 +1,14 @@
 DROP TABLE IF EXISTS issues;
 CREATE TABLE issues (
-    action_index        INTEGER UNSIGNED NOT NULL, -- Unique action index
-    tick_id             INTEGER UNSIGNED,          -- id of record in index_tickers table
+    action_index        BIGINT UNSIGNED NOT NULL, -- Unique action index
+    tick_id             BIGINT UNSIGNED,          -- id of record in index_tickers table
     max_supply          VARCHAR(250),              -- Maximum token supply (1000000000000000000000.000000000000000000 = 40 Characters)
     max_mint            VARCHAR(250),              -- Maximum amount of supply a MINT transaction can issue
     decimals            VARCHAR(2),                -- Number of decimal places token should have (max: 18, default: 0)
     description         VARCHAR(250),              -- URL to a an icon to use for this token (48x48 standard size)
     mint_supply         VARCHAR(250),              -- Maximum amount of supply a MINT transaction can issue
-    transfer_id         INTEGER UNSIGNED,          -- id of record in index_addresses table
-    transfer_supply_id  INTEGER UNSIGNED,          -- id of record in index_addresses table
+    transfer_id         BIGINT UNSIGNED,          -- id of record in index_addresses table
+    transfer_supply_id  BIGINT UNSIGNED,          -- id of record in index_addresses table
     lock_max_supply     VARCHAR(1),                -- Locks MAX_SUPPLY
     lock_mint           VARCHAR(1),                -- Locks MINT
     lock_mint_supply    VARCHAR(1),                -- Locks MINT_SUPPLY
@@ -18,15 +18,15 @@ CREATE TABLE issues (
     lock_sleep          VARCHAR(1),                -- Locks SLEEP
     lock_callback       VARCHAR(1),                -- Locks CALLBACK_BLOCK/TICK/AMOUNT
     callback_block      VARCHAR(15),               -- block_index after which CALLBACK cand be used
-    callback_tick_id    INTEGER UNSIGNED,          -- id of record in index_tickers table
+    callback_tick_id    BIGINT UNSIGNED,          -- id of record in index_tickers table
     callback_amount     VARCHAR(250),              -- AMOUNT users get if CALLBACK
-    allow_list          INTEGER UNSIGNED,          -- action_index of a list from the lists table
-    block_list          INTEGER UNSIGNED,          -- action_index of a list from the lists table
+    allow_list          BIGINT UNSIGNED,          -- action_index of a list from the lists table
+    block_list          BIGINT UNSIGNED,          -- action_index of a list from the lists table
     mint_address_max    VARCHAR(250),              -- Maximum amount of supply an address can MINT
     mint_start_block    VARCHAR(15),               -- block_index when MINT transactions are allowed (begin mint)
     mint_stop_block     VARCHAR(15),               -- BLOCK_INDEX when MINT transactions are NOT allowed (end mint)
-    source_id           INTEGER UNSIGNED,          -- id of record in index_addresses table (address that did DEPLOY)
-    status_id           INTEGER UNSIGNED           -- id of record in index_statuses table
+    source_id           BIGINT UNSIGNED,          -- id of record in index_addresses table (address that did DEPLOY)
+    status_id           BIGINT UNSIGNED           -- id of record in index_statuses table
 ) ENGINE=MyISAM DEFAULT CHARSET=utf8 COLLATE=utf8_general_ci;
 
 CREATE UNIQUE INDEX action_index       ON issues (action_index);
