+++ src/sql/transactions.sql
@@ -2,9 +2,9 @@
 
 DROP TABLE IF EXISTS transactions;
 CREATE TABLE transactions (
-  tx_index    INTEGER UNSIGNED NOT NULL,
-  block_index INTEGER UNSIGNED NOT NULL,
-  tx_hash_id  INTEGER UNSIGNED NOT NULL -- id of record in index_transactions table
+  tx_index    BIGINT UNSIGNED NOT NULL,
+  block_index BIGINT UNSIGNED NOT NULL,
+  tx_hash_id  BIGINT UNSIGNED NOT NULL -- id of record in index_transactions table
 ) ENGINE=MyISAM DEFAULT CHARSET=utf8 COLLATE=utf8_general_ci;
 
 CREATE UNIQUE INDEX tx_index    on transactions (tx_index);
